services/sentiment-analysis/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
ollama==0.4.4
fastapi==0.115.6
uvicorn==0.22.0
prometheus-client==0.17.1
pydantic==2.10.4
redis==5.2.1
numpy==1.26.4
asyncpg==0.29.0